requirements.txt
# Azure Functions
azure-functions==1.18.0

# Azure SDK
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# OpenAI
openai==1.12.0

# Data validation and models
pydantic>=2.10.3
pydantic-settings==2.7.0

# NLP and text processing
nltk==3.8.1
spacy==3.8.3
# spaCy model - must be installed via URL for Azure Functions remote build
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10

# YAML processing for rules
PyYAML==6.0.1

# HTTP and async
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# Reporting and export
reportlab==4.0.9
openpyxl==3.1.2

# Logging and monitoring
applicationinsights==0.11.10
opencensus-ext-azure==1.1.13

# Pin grpc/protobuf to avoid google._upb module cache warnings
grpcio==1.60.0
protobuf==4.25.2
